# examples/shop/openapi.yaml
openapi: 3.0.3
info:
  title: AAT Shop Sandbox API
  version: 1.0.0
  description: |
    Contract for the offline e-commerce sandbox served by `aat-sandbox serve`.
    Two listeners: the shop API (catalog, carts, orders, shipments; OAuth2
    bearer token) and the payments API (charges, refunds; `X-API-Key`).
    Every path is prefixed by a region (`/us/v1`, `/eu/v1`) with independent
    data, currency, tax rules, shipping tiers, and coupons.

    Amounts are integers in minor currency units (cents); `*Display` fields
    carry the formatted value. Errors use one envelope:
    `{"error": {"code": "...", "message": "..."}}`.

    Token endpoint (outside the versioned prefix): `POST /oauth/token` with
    `grant_type=password&username=demo&password=demo&client_id=aat-shop&client_secret=aat-shop-secret`
    (or `grant_type=client_credentials`).
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8765/{region}/v1
    description: Shop API (bearer token)
    variables:
      region:
        enum: [us, eu]
        default: us
        description: us = USD with sales tax added; eu = EUR with VAT included and no overnight tier
...
security:
  - shopToken: []
paths:
...
  /orders/{orderId}/ship:
    parameters:
      - $ref: '#/components/parameters/orderId'
    post:
      operationId: shipOrder
      summary: Ship a paid order
      description: Allowed from `paid` only. Takes about 600 ms.
      tags: [orders, shipments]
      responses:
        '201':
          description: The new shipment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
  /payments/charges:
    servers:
      - $ref: '#/components/x-servers/payments'
    post:
      operationId: paymentCharge
      summary: Charge an order
      description: |
        Served by the payments listener (default port 8766) and authenticated
        with `X-API-Key`. The amount must equal the order total; the order
        moves `created` to `paid`. Card `4000000000000002` is declined (402);
        gift card `GC-10-DEMO` has insufficient funds (402) while
        `GC-100-DEMO` and `GC-500-DEMO` succeed. Takes about 350 ms.
      tags: [payments]
      security:
        - paymentsKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChargeRequest'
      responses:
        '201':
          description: The captured payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/Error'
        '402':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /payments/refunds:
    servers:
      - $ref: '#/components/x-servers/payments'
    post:
      operationId: paymentRefund
      summary: Refund an order's captured payment
      description: Keyed by order id. Full refund unless `amount` is given. 409 when nothing is refundable.
      tags: [payments]
      security:
        - paymentsKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
      responses:
        '201':
          description: The refund
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Refund'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
components:
...
  securitySchemes:
    shopToken:
      type: oauth2
      description: Demo credentials demo/demo, client aat-shop/aat-shop-secret
      flows:
        password:
          tokenUrl: http://localhost:8765/oauth/token
          scopes:
            shop: Full access to the shop API
        clientCredentials:
          tokenUrl: http://localhost:8765/oauth/token
          scopes:
            shop: Full access to the shop API
    paymentsKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Demo key pay-demo-key
  parameters:
...
    orderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Money:
      type: integer
      format: int64
      description: Amount in minor currency units (cents)
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code (for example INVALID_TRANSITION)
            message:
              type: string
...
    Shipment:
      type: object
      required: [shipmentId, orderId, orderStatus, status, carrier, trackingNumber, eta, createdAt]
      properties:
        shipmentId:
          type: string
        orderId:
          type: string
        orderStatus:
          type: string
          enum: [shipped, delivered]
        status:
          type: string
          enum: [in_transit, delivered]
        carrier:
          type: string
        trackingNumber:
          type: string
        eta:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
...
    ChargeRequest:
      type: object
      required: [orderId, amount, currency, method]
      properties:
        orderId:
          type: string
        amount:
          $ref: '#/components/schemas/Money'
        currency:
          type: string
        method:
          type: string
          enum: [card, gift_card, paypal]
        cardNumber:
          type: string
          description: Required for method card; 4000000000000002 is declined
        giftCardCode:
          type: string
          description: Required for method gift_card; GC-10-DEMO, GC-100-DEMO, GC-500-DEMO
        paypalEmail:
          type: string
          description: Required for method paypal
    Payment:
      type: object
      required: [paymentId, orderId, orderStatus, status, method, amount, amountDisplay, currency, createdAt]
      properties:
        paymentId:
          type: string
        orderId:
          type: string
        orderStatus:
          type: string
        status:
          type: string
          enum: [captured, refunded]
        method:
          type: string
          enum: [card, gift_card, paypal]
        amount:
          $ref: '#/components/schemas/Money'
        amountDisplay:
          type: string
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
    RefundRequest:
      type: object
      required: [orderId]
      properties:
        orderId:
          type: string
        amount:
          $ref: '#/components/schemas/Money'
    Refund:
      type: object
      required: [refundId, paymentId, orderId, status, amount, amountDisplay, currency, createdAt]
      properties:
        refundId:
          type: string
        paymentId:
          type: string
        orderId:
          type: string
        status:
          type: string
          enum: [refunded]
        amount:
          $ref: '#/components/schemas/Money'
        amountDisplay:
          type: string
        currency:
          type: string
        createdAt:
          type: string
          format: date-time